Add application service classes.
